rename sdk-config-parser.ts
